[dev] integrated module
